requirements.txt
arcade
scipy
numpy
imageio
pydub
mutagen
python-ffmpeg
Pillow